Changed the variable 'fiftyBanknotes' from int type to double

+++ beginner/uri_1003/src/bee_1018/Banknotes.java
@@ -30,7 +30,7 @@ public class Banknotes {
 
         int restOfOneHundredBanknotes =  banknoteEntered - (oneHundredBanknotes *100);
 
-        int fiftyBanknotes = restOfOneHundredBanknotes / 50 ;
+        double fiftyBanknotes = restOfOneHundredBanknotes / 50 ;
 
 
         double restOfFiftyBanknotes = ((restOfOneHundredBanknotes % 50) - 50);
